+++ src/test/groovy/org/jenkinsci/gradle/plugins/jpi/JpiManifestSpec.groovy
@@ -233,7 +233,7 @@ version = '1.2'"""
                 coreVersion = '1.509.3'
             }
             dependencies {
-                jenkinsPlugins 'org.jenkins-ci.plugins:ant:1.2@jar'
+                jenkinsPlugins 'org.jenkins-ci.plugins:ant:1.2'
             }
         }
         (project as ProjectInternal).evaluate()
@@ -255,8 +255,8 @@ version = '1.2'"""
                 coreVersion = '1.509.3'
             }
             dependencies {
-                jenkinsPlugins 'org.jenkinsci.plugins:git:1.1.15@jar'
-                jenkinsPlugins 'org.jenkins-ci.plugins:ant:1.2@jar'
+                jenkinsPlugins 'org.jenkinsci.plugins:git:1.1.15'
+                jenkinsPlugins 'org.jenkins-ci.plugins:ant:1.2'
             }
         }
         (project as ProjectInternal).evaluate()
@@ -278,7 +278,7 @@ version = '1.2'"""
                 coreVersion = '1.509.3'
             }
             dependencies {
-                optionalJenkinsPlugins 'org.jenkins-ci.plugins:ant:1.2@jar'
+                optionalJenkinsPlugins 'org.jenkins-ci.plugins:ant:1.2'
             }
         }
         (project as ProjectInternal).evaluate()
@@ -300,8 +300,8 @@ version = '1.2'"""
                 coreVersion = '1.509.3'
             }
             dependencies {
-                optionalJenkinsPlugins 'org.jenkinsci.plugins:git:1.1.15@jar'
-                optionalJenkinsPlugins 'org.jenkins-ci.plugins:ant:1.2@jar'
+                optionalJenkinsPlugins 'org.jenkinsci.plugins:git:1.1.15'
+                optionalJenkinsPlugins 'org.jenkins-ci.plugins:ant:1.2'
             }
         }
         (project as ProjectInternal).evaluate()
@@ -323,10 +323,10 @@ version = '1.2'"""
                 coreVersion = '1.509.3'
             }
             dependencies {
-                jenkinsPlugins 'org.jenkinsci.plugins:git:1.1.15@jar'
-                jenkinsPlugins 'org.jenkins-ci.plugins:ant:1.2@jar'
-                optionalJenkinsPlugins 'org.jenkins-ci.plugins:cloudbees-folder:4.2@jar'
-                optionalJenkinsPlugins 'org.jenkins-ci.plugins:credentials:1.9.4@jar'
+                jenkinsPlugins 'org.jenkinsci.plugins:git:1.1.15'
+                jenkinsPlugins 'org.jenkins-ci.plugins:ant:1.2'
+                optionalJenkinsPlugins 'org.jenkins-ci.plugins:cloudbees-folder:4.2'
+                optionalJenkinsPlugins 'org.jenkins-ci.plugins:credentials:1.9.4'
             }
         }
         (project as ProjectInternal).evaluate()