+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 0.24.0 (unreleased)
+
 ## 0.23.0 (2017-10-10)
 
   * updated to Gradle 3.5.1